features/requirements.txt
behave>=1.2.6
PyMySQL>=0.9.2
simplejson>=3.11.1
requests>=2.18.4
PyHamcrest>=1.9.0
mysql-replication==0.13
wheel>=0.29.0
arrow>=0.12.1
psycopg2-binary>=2.8.4